rename it, apply it to channels instead of messages
